backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Redis and job queue
redis==5.0.1
rq==1.15.1

# HTTP client
httpx==0.25.2

# Authentication and security
pyjwt[crypto]==2.8.0
cryptography==41.0.7

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
responses==0.24.1

# Utilities
python-dotenv==1.0.0

